requirements.txt
my-sql-connector
Flask